services/analytics/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyJWT==2.9.0
httpx==0.27.2
